refactor(svelte2tsx): バリデーションとソース走査を validation.rs・utils/ へ
- `validation.rs`: 公式 svelte2tsx が svelte のパーサ側で弾く検査の再現（rsvelte 固有）
- `utils/htmlxparser.rs`: `<style>` 内容のブランキングと孤立 `<script>` の回収
- `utils/lexical.rs`: 語境界 / 識別子スキャン

`is_ident_char` と `lexical_identifiers` 系は複数の呼び出し元から使われる共有ヘルパーなので
`utils/lexical.rs` に集約した。

純粋な移動でロジック変更はなく、出力はバイト単位で不変。
