+++ crates/rsvelte_core/src/svelte2tsx/svelte2tsx.rs
@@ -11,6 +11,11 @@ use crate::compiler::phases::phase1_parse::{self, ParseOptions};
 use super::magic_string::{GenerateMapOptions, MagicString};
 use super::script::{ComponentEvents, ExportedNames};
 use super::template;
+use super::utils::htmlxparser::{blank_style_content, blank_style_tags, remove_orphan_scripts};
+use super::utils::lexical::{
+    contains_word, is_ident_char, lexical_identifiers, lexical_identifiers_in_expressions,
+};
+use super::validation::{validate_debug_tag_arguments, validate_meta_element_placement};
 
 pub use super::interfaces::{
     RewriteExternalImportsOptions, Svelte2TsxMode, Svelte2TsxNamespace, Svelte2TsxOptions,
@@ -25,311 +30,6 @@ pub use super::utils::error::Svelte2TsxError;
 #[inline]
 pub(crate) fn slice_src(source: &str, start: usize, end: usize) -> &str {
     source.get(start..end).unwrap_or("")
-}
-
-/// Case-insensitive byte search: position of the first occurrence of
-/// `needle` in `haystack[from..]` (absolute index). ASCII-only folding —
-/// exactly what `to_ascii_lowercase` matching gave, without allocating a
-/// lowercased copy of the whole source.
-fn find_ci(haystack: &[u8], from: usize, needle: &[u8]) -> Option<usize> {
-    if from > haystack.len() {
-        return None;
-    }
-    haystack[from..]
-        .windows(needle.len())
-        .position(|w| w.eq_ignore_ascii_case(needle))
-        .map(|p| from + p)
-}
-
-/// Replace the content of every `<style …>…</style>` with spaces (newlines and
-/// carriage returns preserved) so the parser never CSS-parses it. Works at the
-/// BYTE level so the result is exactly the same length as `source` — every AST
-/// offset still indexes the original source. Case-insensitive on the tag name.
-fn blank_style_content(source: &str) -> String {
-    let mut bytes = source.as_bytes().to_vec();
-    let sb = source.as_bytes();
-    let mut search = 0usize;
-    while let Some(tag_start) = find_ci(sb, search, b"<style") {
-        // Must be the `<style` element, not e.g. `<styled`.
-        let after = sb.get(tag_start + 6).copied();
-        if !matches!(
-            after,
-            Some(b'>') | Some(b' ') | Some(b'\t') | Some(b'\n') | Some(b'\r') | Some(b'/') | None
-        ) {
-            search = tag_start + 6;
-            continue;
-        }
-        let Some(gt) = find_ci(sb, tag_start, b">") else {
-            break;
-        };
-        let content_start = gt + 1;
-        // Self-closing `<style/>` → no content to blank.
-        if content_start >= 2 && sb[content_start - 2] == b'/' {
-            search = content_start;
-            continue;
-        }
-        let Some(content_end) = find_ci(sb, content_start, b"</style") else {
-            break;
-        };
-        for b in &mut bytes[content_start..content_end] {
-            if *b != b'\n' && *b != b'\r' {
-                *b = b' ';
-            }
-        }
-        search = content_end;
-    }
-    String::from_utf8(bytes).unwrap_or_else(|_| source.to_string())
-}
-
-/// Validate that every `{@debug …}` argument is a plain identifier, returning a
-/// template error otherwise — mirrors svelte's parse-time check (rsvelte's lives
-/// in the analyze DebugTag visitor, which svelte2tsx doesn't run).
-fn validate_debug_tag_arguments(ast: &Root, source: &str) -> Result<(), Svelte2TsxError> {
-    use crate::ast::template::{Fragment, TemplateNode as N};
-
-    fn arg_is_identifier(expr: &crate::ast::js::Expression, source: &str) -> bool {
-        match expr.node_type() {
-            Some("Identifier") => true,
-            Some(_) => false,
-            // Lazy/unresolved expression: accept only a bare identifier token.
-            None => match (expr.start(), expr.end()) {
-                (Some(s), Some(e))
-                    if (s as usize) < (e as usize) && (e as usize) <= source.len() =>
-                {
-                    let t = slice_src(source, s as usize, e as usize).trim();
-                    let mut chars = t.chars();
-                    match chars.next() {
-                        Some(c0) if c0.is_alphabetic() || c0 == '_' || c0 == '$' => {
-                            chars.all(|c| c.is_alphanumeric() || c == '_' || c == '$')
-                        }
-                        _ => false,
-                    }
-                }
-                _ => false,
-            },
-        }
-    }
-
-    fn check_fragment(frag: &Fragment, source: &str) -> Result<(), Svelte2TsxError> {
-        for node in &frag.nodes {
-            check_node(node, source)?;
-        }
-        Ok(())
-    }
-
-    fn check_node(node: &N, source: &str) -> Result<(), Svelte2TsxError> {
-        match node {
-            N::DebugTag(tag) => {
-                for id in &tag.identifiers {
-                    if !arg_is_identifier(id, source) {
-                        return Err(Svelte2TsxError::Template(
-                            "{@debug ...} arguments must be identifiers, not arbitrary expressions"
-                                .to_string(),
-                        ));
-                    }
-                }
-            }
-            N::RegularElement(e) => check_fragment(&e.fragment, source)?,
-            N::Component(c) => check_fragment(&c.fragment, source)?,
-            N::SvelteComponent(c) => check_fragment(&c.fragment, source)?,
-            N::SvelteElement(e) => check_fragment(&e.fragment, source)?,
-            N::SvelteHead(e)
-            | N::SvelteFragment(e)
-            | N::SvelteBody(e)
-            | N::SvelteWindow(e)
-            | N::SvelteDocument(e)
-            | N::SvelteBoundary(e)
-            | N::SvelteOptions(e)
-            | N::SvelteSelf(e) => check_fragment(&e.fragment, source)?,
-            N::TitleElement(e) => check_fragment(&e.fragment, source)?,
-            N::SlotElement(e) => check_fragment(&e.fragment, source)?,
-            N::IfBlock(b) => {
-                check_fragment(&b.consequent, source)?;
-                if let Some(alt) = &b.alternate {
-                    check_fragment(alt, source)?;
-                }
-            }
-            N::EachBlock(b) => {
-                check_fragment(&b.body, source)?;
-                if let Some(fb) = &b.fallback {
-                    check_fragment(fb, source)?;
-                }
-            }
-            N::KeyBlock(b) => check_fragment(&b.fragment, source)?,
-            N::SnippetBlock(b) => check_fragment(&b.body, source)?,
-            N::AwaitBlock(b) => {
-                if let Some(f) = &b.pending {
-                    check_fragment(f, source)?;
-                }
-                if let Some(f) = &b.then {
-                    check_fragment(f, source)?;
-                }
-                if let Some(f) = &b.catch {
-                    check_fragment(f, source)?;
-                }
-            }
-            _ => {}
-        }
-        Ok(())
-    }
-
-    check_fragment(&ast.fragment, source)
-}
-
-/// True when a component carries a `use:` action directive. svelte rejects ALL
-/// of `use:`/`transition:`/`animate:`/`class:`/`style:` on a component
-/// (`component_invalid_directive`) at 2-analyze, but official **svelte2tsx**
-/// skips analyze and actually LOWERS class/style/transition/animate on a
-/// component (to `ensureType`/`ensureTransition` suffixes) — only `use:` makes
-/// it CRASH (`element.addAction is not a function`). So, for error-parity with
-/// svelte2tsx specifically, only `use:` triggers an error here.
-fn component_has_invalid_directive(attributes: &[crate::ast::Attribute<'_>]) -> bool {
-    use crate::ast::Attribute as A;
-    attributes.iter().any(|a| matches!(a, A::UseDirective(_)))
-}
-
-/// Validate `<svelte:window/body/document/head/options>` placement and
-/// uniqueness, mirroring svelte's PARSE-time `svelte_meta_duplicate` /
-/// `svelte_meta_invalid_placement` checks (1-parse/state/element.js). rsvelte's
-/// compiler defers these to phase-2 analysis, which svelte2tsx skips — but
-/// official svelte2tsx parses with svelte and so rejects these at parse. Each
-/// of these five "root-only meta tags" must appear at most once and only as a
-/// direct child of the component root (not inside any element or block).
-fn validate_meta_element_placement(ast: &Root<'_>, source: &str) -> Result<(), Svelte2TsxError> {
-    use crate::ast::template::{Fragment, TemplateNode as N};
-    use std::collections::HashSet;
-
-    // `<svelte:element>` requires a `this` attribute with a value. svelte's
-    // parser stores it as the element's `tag` expression; a missing / valueless
-    // `this` leaves an empty span. Official svelte2tsx rejects it.
-    fn dynamic_element_tag_is_empty(tag: &crate::ast::js::Expression<'_>, source: &str) -> bool {
-        match (tag.start(), tag.end()) {
-            (Some(s), Some(e)) if (s as usize) < (e as usize) && (e as usize) <= source.len() => {
-                slice_src(source, s as usize, e as usize).trim().is_empty()
-            }
-            _ => true,
-        }
-    }
-
-    fn meta_name<'b>(node: &'b N<'_>) -> Option<&'b str> {
-        match node {
-            N::SvelteWindow(e)
-            | N::SvelteBody(e)
-            | N::SvelteDocument(e)
-            | N::SvelteHead(e)
-            | N::SvelteOptions(e) => Some(e.name.as_str()),
-            _ => None,
-        }
-    }
-
-    fn check_fragment(
-        frag: &Fragment,
-        at_root: bool,
-        seen: &mut HashSet<String>,
-        source: &str,
-    ) -> Result<(), Svelte2TsxError> {
-        for node in &frag.nodes {
-            check_node(node, at_root, seen, source)?;
-        }
-        Ok(())
-    }
-
-    fn check_node(
-        node: &N,
-        at_root: bool,
-        seen: &mut HashSet<String>,
-        source: &str,
-    ) -> Result<(), Svelte2TsxError> {
-        if let N::SvelteElement(e) = node
-            && dynamic_element_tag_is_empty(&e.tag, source)
-        {
-            return Err(Svelte2TsxError::Template(
-                "`<svelte:element>` must have a 'this' attribute with a value".to_string(),
-            ));
-        }
-        if let Some(name) = meta_name(node) {
-            if !at_root {
-                return Err(Svelte2TsxError::Template(format!(
-                    "`<{}>` tags cannot be inside elements or blocks",
-                    name
-                )));
-            }
-            if !seen.insert(name.to_string()) {
-                return Err(Svelte2TsxError::Template(format!(
-                    "A component can only have one `<{}>` element",
-                    name
-                )));
-            }
-        }
-        // `use:` / `transition:` / `in:` / `out:` / `animate:` / `class:` /
-        // `style:` directives are not valid on a component — svelte raises
-        // `component_invalid_directive` (2-analyze). Official svelte2tsx skips
-        // analyze and instead CRASHES on these (`element.addAction is not a
-        // function`); either way it produces an error, so raise one here for
-        // error-parity. (`bind:` / `on:` / `let:` / spread / `@attach` are OK.)
-        if let N::Component(c) = node
-            && component_has_invalid_directive(&c.attributes)
-        {
-            return Err(Svelte2TsxError::Template(
-                "This type of directive is not valid on components".to_string(),
-            ));
-        }
-        if let N::SvelteComponent(c) = node
-            && component_has_invalid_directive(&c.attributes)
-        {
-            return Err(Svelte2TsxError::Template(
-                "This type of directive is not valid on components".to_string(),
-            ));
-        }
-        // Recurse into children — anything nested below this node is no longer
-        // at root level.
-        match node {
-            N::RegularElement(e) => check_fragment(&e.fragment, false, seen, source)?,
-            N::Component(c) => check_fragment(&c.fragment, false, seen, source)?,
-            N::SvelteComponent(c) => check_fragment(&c.fragment, false, seen, source)?,
-            N::SvelteElement(e) => check_fragment(&e.fragment, false, seen, source)?,
-            N::SvelteHead(e)
-            | N::SvelteFragment(e)
-            | N::SvelteBody(e)
-            | N::SvelteWindow(e)
-            | N::SvelteDocument(e)
-            | N::SvelteBoundary(e)
-            | N::SvelteOptions(e)
-            | N::SvelteSelf(e) => check_fragment(&e.fragment, false, seen, source)?,
-            N::TitleElement(e) => check_fragment(&e.fragment, false, seen, source)?,
-            N::SlotElement(e) => check_fragment(&e.fragment, false, seen, source)?,
-            N::IfBlock(b) => {
-                check_fragment(&b.consequent, false, seen, source)?;
-                if let Some(alt) = &b.alternate {
-                    check_fragment(alt, false, seen, source)?;
-                }
-            }
-            N::EachBlock(b) => {
-                check_fragment(&b.body, false, seen, source)?;
-                if let Some(fb) = &b.fallback {
-                    check_fragment(fb, false, seen, source)?;
-                }
-            }
-            N::KeyBlock(b) => check_fragment(&b.fragment, false, seen, source)?,
-            N::SnippetBlock(b) => check_fragment(&b.body, false, seen, source)?,
-            N::AwaitBlock(b) => {
-                if let Some(f) = &b.pending {
-                    check_fragment(f, false, seen, source)?;
-                }
-                if let Some(f) = &b.then {
-                    check_fragment(f, false, seen, source)?;
-                }
-                if let Some(f) = &b.catch {
-                    check_fragment(f, false, seen, source)?;
-                }
-            }
-            _ => {}
-        }
-        Ok(())
-    }
-
-    let mut seen = HashSet::new();
-    check_fragment(&ast.fragment, true, &mut seen, source)
 }
 
 /// Convert a Svelte component source to TypeScript/TSX for type checking.
@@ -2019,28 +1719,6 @@ pub fn svelte2tsx(
     })
 }
 
-/// Remove embedded `<script>` tags that are NOT the top-level instance / module
-/// script (they sit inside attribute values or template-literal expressions).
-/// Overwriting their range with `""` truncates any attribute whose source span
-/// covers the range; the joined content is returned for injection into the
-/// `$$render()` body when the file has no top-level script.
-fn remove_orphan_scripts(ast: &Root, source: &str, str: &mut MagicString) -> String {
-    let orphan_scripts = find_orphan_scripts(ast, source);
-    // Remove orphan scripts from the MagicString (must happen BEFORE
-    // process_template_inplace so the overwrite is in place when the template
-    // emits Seg::Src ranges that span the orphan range).
-    for &(s, e, _) in &orphan_scripts {
-        str.overwrite(s, e, "");
-    }
-    // Collect content for injection into $$render() body. Only matters when
-    // there is no top-level instance/module script (the "no script" path below).
-    orphan_scripts
-        .iter()
-        .map(|(_, _, content)| content.as_str())
-        .collect::<Vec<_>>()
-        .join("\n")
-}
-
 /// Emit the `<svelte:options>` tag as a `svelteHTML.createElement(...)` call.
 /// The parser stores svelte:options in `ast.options` (not in `fragment.nodes`),
 /// so it is handled separately.
@@ -2132,117 +1810,6 @@ fn emit_svelte_options_element(ast: &Root, source: &str, str: &mut MagicString) 
         attrs_str
     );
     str.overwrite(options_node.start, options_node.end, &replacement);
-}
-
-/// Blank out `<style>` tags (CSS is not relevant for TSX type checking). First
-/// blanks any style tag the parser captured in `ast.css`, then always runs a
-/// fallback scanner to catch style tags the parser did not capture (e.g.,
-/// `<style global>`, `<style lang="...">`).
-fn blank_style_tags(ast: &Root, source: &str, str: &mut MagicString) {
-    let mut blanked_style_ranges: Vec<(usize, usize)> = Vec::new();
-    if let Some(ref css) = ast.css
-        && css.start < css.end
-    {
-        // Only blank the CSS range when the close tag is well-formed (exact
-        // `</style>` with no whitespace before `>`). When the close tag is
-        // malformed (e.g. `</style   >`), the official svelte2tsx regex does
-        // not match the style tag and it is left as raw text in the output.
-        // Mirror that: skip blanking so the raw `<style>…</style   >` text
-        // appears verbatim in the async template body.
-        let has_proper_style_close = {
-            let slice = slice_src(source, css.start as usize, css.end as usize);
-            slice
-                .as_bytes()
-                .windows(8)
-                .any(|w| w.eq_ignore_ascii_case(b"</style>"))
-        };
-        if has_proper_style_close {
-            // Also blank any trailing whitespace after the style tag
-            let mut blank_end = css.end;
-            let bytes = source.as_bytes();
-            while (blank_end as usize) < bytes.len() {
-                let b = bytes[blank_end as usize];
-                if b == b' ' || b == b'\t' || b == b'\n' || b == b'\r' {
-                    blank_end += 1;
-                } else {
-                    break;
-                }
-            }
-            str.overwrite(css.start, blank_end, "");
-            blanked_style_ranges.push((css.start as usize, blank_end as usize));
-        }
-    }
-    {
-        // Fallback: scan source for <style tags that the parser didn't
-        // capture in ast.css (e.g., <style global>, <style lang="...">).
-        // Blank them out by finding the matching </style>.
-        // Exclude positions inside script tags to avoid matching <style>
-        // inside template literals or string content.
-        let script_ranges: Vec<(usize, usize)> = {
-            let mut ranges = Vec::new();
-            if let Some(ref inst) = ast.instance {
-                ranges.push((inst.start as usize, inst.end as usize));
-            }
-            if let Some(ref module) = ast.module {
-                ranges.push((module.start as usize, module.end as usize));
-            }
-            ranges
-        };
-        let is_inside_script =
-            |pos: usize| -> bool { script_ranges.iter().any(|&(s, e)| pos >= s && pos < e) };
-        let is_already_blanked = |pos: usize| -> bool {
-            blanked_style_ranges
-                .iter()
-                .any(|&(s, e)| pos >= s && pos < e)
-        };
-
-        // Direct case-sensitive substring search over the original source.
-        // The previous implementation called `source.to_lowercase()` once
-        // per call, allocating a full copy of the source for case-
-        // insensitive matching. Svelte HTML is lowercase in practice
-        // (the parser only recognises lowercase tags), so the lowercase
-        // copy is unnecessary overhead.
-        let bytes = source.as_bytes();
-        let mut search_from = 0;
-        while let Some(rel) = source[search_from..].find("<style") {
-            let abs_start = search_from + rel;
-            if is_inside_script(abs_start) {
-                search_from = abs_start + 1;
-                continue;
-            }
-            if is_already_blanked(abs_start) {
-                search_from = abs_start + 1;
-                continue;
-            }
-            let after_tag = abs_start + 6;
-            if after_tag < bytes.len() {
-                let next_ch = bytes[after_tag];
-                if (next_ch == b' '
-                    || next_ch == b'>'
-                    || next_ch == b'\n'
-                    || next_ch == b'\r'
-                    || next_ch == b'\t'
-                    || next_ch == b'/')
-                    && let Some(close_off) = source[abs_start..].find("</style>")
-                {
-                    let abs_end = abs_start + close_off + 8; // 8 = len("</style>")
-                    let mut blank_end = abs_end as u32;
-                    while (blank_end as usize) < bytes.len() {
-                        let b = bytes[blank_end as usize];
-                        if b == b' ' || b == b'\t' || b == b'\n' || b == b'\r' {
-                            blank_end += 1;
-                        } else {
-                            break;
-                        }
-                    }
-                    str.overwrite(abs_start as u32, blank_end, "");
-                    search_from = abs_end;
-                    continue;
-                }
-            }
-            search_from = abs_start + 1;
-        }
-    }
 }
 
 /// Analyze and relocate top-level `{#snippet}` blocks. Non-hoistable snippets
@@ -2794,46 +2361,6 @@ fn extract_component_documentation(fragment: &crate::ast::template::Fragment) ->
     None
 }
 
-/// Conservative whole-word substring search. Returns `true` when `needle`
-/// appears in `haystack` with non-identifier bytes on either side. Used as
-/// a fast pre-filter before an expensive AST parse — false positives waste
-/// a few microseconds, but false negatives must be impossible, which holds
-/// because any real `import` or `await` statement contains those exact
-/// bytes as a word.
-fn contains_word(haystack: &[u8], needle: &[u8]) -> bool {
-    if needle.is_empty() || haystack.len() < needle.len() {
-        return false;
-    }
-    let first = needle[0];
-    let mut i = 0;
-    while i + needle.len() <= haystack.len() {
-        let off = match memchr::memchr(first, &haystack[i..]) {
-            Some(o) => o,
-            None => return false,
-        };
-        let pos = i + off;
-        if pos + needle.len() > haystack.len() {
-            return false;
-        }
-        if &haystack[pos..pos + needle.len()] == needle {
-            let before_ok = pos == 0
-                || !(haystack[pos - 1].is_ascii_alphanumeric()
-                    || haystack[pos - 1] == b'_'
-                    || haystack[pos - 1] == b'$');
-            let after = pos + needle.len();
-            let after_ok = after == haystack.len()
-                || !(haystack[after].is_ascii_alphanumeric()
-                    || haystack[after] == b'_'
-                    || haystack[after] == b'$');
-            if before_ok && after_ok {
-                return true;
-            }
-        }
-        i = pos + 1;
-    }
-    false
-}
-
 /// Find the start of `</script>` tag by scanning backwards from the script end position.
 fn find_script_close_tag_start(source: &str, script_end: u32) -> u32 {
     let bytes = source.as_bytes();
@@ -3292,139 +2819,6 @@ fn is_snippet_module_hoistable(
     true
 }
 
-/// Lex a string into ASCII-identifier tokens. Skips `//` and `/* */` comments
-/// and `'`, `"`, ``\``` strings so identifiers inside literals aren't picked
-/// up as references.
-fn lexical_identifiers(text: &str) -> Vec<String> {
-    let mut out = Vec::new();
-    let bytes = text.as_bytes();
-    let len = bytes.len();
-    let mut i = 0usize;
-    while i < len {
-        let b = bytes[i];
-        if b == b'/' && i + 1 < len {
-            if bytes[i + 1] == b'/' {
-                while i < len && bytes[i] != b'\n' {
-                    i += 1;
-                }
-                continue;
-            } else if bytes[i + 1] == b'*' {
-                i += 2;
-                while i + 1 < len && !(bytes[i] == b'*' && bytes[i + 1] == b'/') {
-                    i += 1;
-                }
-                i = (i + 2).min(len);
-                continue;
-            }
-        }
-        if b == b'\'' || b == b'"' || b == b'`' {
-            let quote = b;
-            i += 1;
-            while i < len && bytes[i] != quote {
-                if bytes[i] == b'\\' && i + 1 < len {
-                    i += 2;
-                    continue;
-                }
-                i += 1;
-            }
-            i = (i + 1).min(len);
-            continue;
-        }
-        if is_ident_char(b) && !b.is_ascii_digit() {
-            let start = i;
-            i += 1;
-            while i < len && is_ident_char(bytes[i]) {
-                i += 1;
-            }
-            out.push(text[start..i].to_string());
-            continue;
-        }
-        i += 1;
-    }
-    out
-}
-
-/// Extract identifiers that appear inside Svelte template expression blocks
-/// (`{...}` / `{#...}` / `{:...}` / `{/...}` / `{@...}`).
-///
-/// This is intentionally more conservative than `lexical_identifiers`: it only
-/// yields tokens found inside the `{ … }` delimiters of template tags, which
-/// is roughly what periscopic's JS AST scope analysis would see as free
-/// variables.  HTML attribute NAMES (e.g. the `state` in `data-state`) are
-/// outside these delimiters and therefore not returned, preventing false
-/// positives when checking `instance_script_loose_dollar_names`.
-fn lexical_identifiers_in_expressions(text: &str) -> Vec<String> {
-    let bytes = text.as_bytes();
-    let len = bytes.len();
-    let mut out = Vec::new();
-    let mut i = 0usize;
-
-    while i < len {
-        if bytes[i] != b'{' {
-            i += 1;
-            continue;
-        }
-        // Found `{` — scan the delimited expression region.
-        i += 1; // skip `{`
-        let mut depth = 1usize;
-        let expr_start = i;
-        while i < len && depth > 0 {
-            let b = bytes[i];
-            match b {
-                b'{' => {
-                    depth += 1;
-                    i += 1;
-                }
-                b'}' => {
-                    depth -= 1;
-                    i += 1;
-                }
-                b'\'' | b'"' | b'`' => {
-                    let q = b;
-                    i += 1;
-                    while i < len && bytes[i] != q {
-                        if bytes[i] == b'\\' && i + 1 < len {
-                            i += 2;
-                        } else {
-                            i += 1;
-                        }
-                    }
-                    if i < len {
-                        i += 1;
-                    }
-                }
-                b'/' if i + 1 < len && bytes[i + 1] == b'/' => {
-                    while i < len && bytes[i] != b'\n' {
-                        i += 1;
-                    }
-                }
-                b'/' if i + 1 < len && bytes[i + 1] == b'*' => {
-                    i += 2;
-                    while i + 1 < len && !(bytes[i] == b'*' && bytes[i + 1] == b'/') {
-                        i += 1;
-                    }
-                    i = (i + 2).min(len);
-                }
-                _ => {
-                    i += 1;
-                }
-            }
-        }
-        let expr_end = i.saturating_sub(1); // don't include the closing `}`
-        // `text.get(..)` rather than direct slicing: on a (parser-rejected)
-        // unterminated `{` whose body ends mid-multibyte-char, `expr_end` could
-        // land off a char boundary — fall through instead of panicking.
-        if expr_start < expr_end
-            && let Some(expr_text) = text.get(expr_start..expr_end)
-        {
-            for tok in lexical_identifiers(expr_text) {
-                out.push(tok);
-            }
-        }
-    }
-    out
-}
-
 /// Return true if `type_text` mentions any of `names` as a whole identifier
 /// (i.e. surrounded by non-identifier characters on both sides).
 ///
@@ -3456,11 +2850,6 @@ fn type_text_references_any(type_text: &str, names: &std::collections::HashSet<S
         }
     }
     false
-}
-
-#[inline]
-fn is_ident_char(b: u8) -> bool {
-    b.is_ascii_alphanumeric() || b == b'_' || b == b'$'
 }
 
 /// Return true if `type_text` contains a `typeof X` value-query whose root
@@ -4937,222 +4326,6 @@ fn lazy_slice_references_rune_global(slice: &str) -> bool {
     false
 }
 
-// =============================================================================
-// Orphan-script detection (Step 7.48)
-// =============================================================================
-
-/// Collect start positions of every `RegularElement` named `"script"` anywhere
-/// in the fragment tree (including inside `{#if}`, `{#each}`, `<svelte:head>`,
-/// nested elements, etc.).
-fn collect_script_element_starts(
-    fragment: &crate::ast::template::Fragment,
-    out: &mut std::collections::HashSet<u32>,
-) {
-    use crate::ast::template::TemplateNode as N;
-    for node in &fragment.nodes {
-        match node {
-            N::RegularElement(e) => {
-                if e.name == "script" {
-                    out.insert(e.start);
-                }
-                collect_script_element_starts(&e.fragment, out);
-            }
-            N::Component(c) => collect_script_element_starts(&c.fragment, out),
-            N::SvelteComponent(c) => collect_script_element_starts(&c.fragment, out),
-            N::SvelteElement(e) => collect_script_element_starts(&e.fragment, out),
-            N::TitleElement(e) => collect_script_element_starts(&e.fragment, out),
-            N::SlotElement(e) => collect_script_element_starts(&e.fragment, out),
-            N::SvelteHead(e)
-            | N::SvelteFragment(e)
-            | N::SvelteBody(e)
-            | N::SvelteWindow(e)
-            | N::SvelteDocument(e)
-            | N::SvelteBoundary(e)
-            | N::SvelteOptions(e)
-            | N::SvelteSelf(e) => collect_script_element_starts(&e.fragment, out),
-            N::IfBlock(b) => {
-                collect_script_element_starts(&b.consequent, out);
-                if let Some(alt) = &b.alternate {
-                    collect_script_element_starts(alt, out);
-                }
-            }
-            N::EachBlock(b) => {
-                collect_script_element_starts(&b.body, out);
-                if let Some(fb) = &b.fallback {
-                    collect_script_element_starts(fb, out);
-                }
-            }
-            N::KeyBlock(b) => collect_script_element_starts(&b.fragment, out),
-            N::SnippetBlock(b) => collect_script_element_starts(&b.body, out),
-            N::AwaitBlock(b) => {
-                if let Some(f) = &b.pending {
-                    collect_script_element_starts(f, out);
-                }
-                if let Some(f) = &b.then {
-                    collect_script_element_starts(f, out);
-                }
-                if let Some(f) = &b.catch {
-                    collect_script_element_starts(f, out);
-                }
-            }
-            _ => {}
-        }
-    }
-}
-
-/// Collect (start, end) ranges of every `HtmlTag` (`{@html}`) node anywhere
-/// in the fragment tree. A `<script>` inside a HtmlTag expression is NOT an
-/// orphan — it's already handled by the `{@html}` output.
-fn collect_html_tag_ranges(fragment: &crate::ast::template::Fragment, out: &mut Vec<(u32, u32)>) {
-    use crate::ast::template::TemplateNode as N;
-    for node in &fragment.nodes {
-        match node {
-            N::HtmlTag(h) => {
-                out.push((h.start, h.end));
-            }
-            N::RegularElement(e) => collect_html_tag_ranges(&e.fragment, out),
-            N::Component(c) => collect_html_tag_ranges(&c.fragment, out),
-            N::SvelteComponent(c) => collect_html_tag_ranges(&c.fragment, out),
-            N::SvelteElement(e) => collect_html_tag_ranges(&e.fragment, out),
-            N::TitleElement(e) => collect_html_tag_ranges(&e.fragment, out),
-            N::SlotElement(e) => collect_html_tag_ranges(&e.fragment, out),
-            N::SvelteHead(e)
-            | N::SvelteFragment(e)
-            | N::SvelteBody(e)
-            | N::SvelteWindow(e)
-            | N::SvelteDocument(e)
-            | N::SvelteBoundary(e)
-            | N::SvelteOptions(e)
-            | N::SvelteSelf(e) => collect_html_tag_ranges(&e.fragment, out),
-            N::IfBlock(b) => {
-                collect_html_tag_ranges(&b.consequent, out);
-                if let Some(alt) = &b.alternate {
-                    collect_html_tag_ranges(alt, out);
-                }
-            }
-            N::EachBlock(b) => {
-                collect_html_tag_ranges(&b.body, out);
-                if let Some(fb) = &b.fallback {
-                    collect_html_tag_ranges(fb, out);
-                }
-            }
-            N::KeyBlock(b) => collect_html_tag_ranges(&b.fragment, out),
-            N::SnippetBlock(b) => collect_html_tag_ranges(&b.body, out),
-            N::AwaitBlock(b) => {
-                if let Some(f) = &b.pending {
-                    collect_html_tag_ranges(f, out);
-                }
-                if let Some(f) = &b.then {
-                    collect_html_tag_ranges(f, out);
-                }
-                if let Some(f) = &b.catch {
-                    collect_html_tag_ranges(f, out);
-                }
-            }
-            _ => {}
-        }
-    }
-}
-
-/// Find "orphan" `<script>…</script>` occurrences in `source` that the Svelte
-/// parser did NOT recognise as a real Script (instance/module) or as a
-/// `RegularElement` named `"script"`. These are typically `<script>` tags
-/// embedded inside attribute values (e.g. `href="</noscript><script>…"`) or
-/// inside template-literal expressions that the HTML parser terminated early.
-///
-/// Returns a list of `(start, end, inner_content)` triples, where `start`/`end`
-/// are byte offsets in `source` and `inner_content` is the raw text between
-/// `<script…>` and `</script>`.
-fn find_orphan_scripts(ast: &Root, source: &str) -> Vec<(u32, u32, String)> {
-    // 1. Collect known "legitimate" script start positions and their full ranges.
-    let mut known_starts: std::collections::HashSet<u32> = std::collections::HashSet::default();
-    // Also collect ranges of instance/module scripts so we can skip `<script>`
-    // occurrences inside their string literals / template content.
-    let mut known_ranges: Vec<(u32, u32)> = Vec::new();
-    if let Some(inst) = &ast.instance {
-        known_starts.insert(inst.start);
-        known_ranges.push((inst.start, inst.end));
-    }
-    if let Some(module) = &ast.module {
-        known_starts.insert(module.start);
-        known_ranges.push((module.start, module.end));
-    }
-    collect_script_element_starts(&ast.fragment, &mut known_starts);
-
-    // 2. Collect HtmlTag ranges — a <script> inside {@html …} is not orphan.
-    let mut html_tag_ranges: Vec<(u32, u32)> = Vec::new();
-    collect_html_tag_ranges(&ast.fragment, &mut html_tag_ranges);
-
-    // 3. Scan the source for `<script` occurrences (case-insensitive, without
-    // allocating a lowercased copy of the whole source).
-    let bytes = source.as_bytes();
-    let mut result: Vec<(u32, u32, String)> = Vec::new();
-    let mut search: usize = 0;
-
-    while search < source.len() {
-        let Some(abs) = find_ci(bytes, search, b"<script") else {
-            break;
-        };
-        let tag_start = abs as u32;
-
-        // Require a proper tag boundary after `<script` (6 bytes for "script").
-        let after_pos = tag_start as usize + 7; // skip '<' + "script"
-        let next_byte = bytes.get(after_pos).copied();
-        if !matches!(
-            next_byte,
-            Some(b'>') | Some(b' ') | Some(b'\t') | Some(b'\n') | Some(b'\r') | Some(b'/') | None
-        ) {
-            search = tag_start as usize + 7;
-            continue;
-        }
-
-        // Skip if this is a recognised script/element node (exact start match).
-        if known_starts.contains(&tag_start) {
-            search = tag_start as usize + 7;
-            continue;
-        }
-
-        // Skip if the tag start falls INSIDE an instance/module script range
-        // (e.g. `<script>` text inside a string literal in the script body).
-        if known_ranges
-            .iter()
-            .any(|&(s, e)| tag_start > s && tag_start < e)
-        {
-            search = tag_start as usize + 7;
-            continue;
-        }
-
-        // Skip if the tag start falls inside a {@html …} range.
-        if html_tag_ranges
-            .iter()
-            .any(|&(s, e)| tag_start > s && tag_start < e)
-        {
-            search = tag_start as usize + 7;
-            continue;
-        }
-
-        // Find the matching `</script>` (case-insensitive).
-        let Some(close_abs) = find_ci(bytes, tag_start as usize, b"</script>") else {
-            break; // unterminated — skip
-        };
-        let close_rel = close_abs - tag_start as usize;
-        let tag_end = tag_start + close_rel as u32 + 9; // 9 = len("</script>")
-
-        // Extract the inner content: everything between `>` of the open tag and
-        // `<` of `</script>`.
-        let open_gt = slice_src(source, tag_start as usize, tag_end as usize)
-            .find('>')
-            .map(|p| tag_start as usize + p + 1)
-            .unwrap_or(tag_start as usize + 8); // fallback: after "<script>"
-        let inner = source[open_gt..tag_start as usize + close_rel].to_string();
-
-        result.push((tag_start, tag_end, inner));
-        search = tag_end as usize;
-    }
-
-    result
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -5193,38 +4366,6 @@ mod tests {
             &set(&[]),
             &set(&["ModuleThing"]),
         ));
-    }
-
-    #[test]
-    fn lexical_identifiers_in_expressions_only_reads_brace_blocks() {
-        // Identifiers inside `{...}` are returned; HTML attribute names like
-        // `data-state` (outside braces) are not — avoids false hoist-blocks.
-        let got = lexical_identifiers_in_expressions("<div data-state={open}>{value}</div>");
-        assert!(got.contains(&"open".to_string()), "{got:?}");
-        assert!(got.contains(&"value".to_string()), "{got:?}");
-        assert!(
-            !got.contains(&"state".to_string()),
-            "attr name skipped: {got:?}"
-        );
-        assert!(
-            !got.contains(&"data".to_string()),
-            "attr name skipped: {got:?}"
-        );
-        // Strings inside an expression are skipped; nested braces are balanced.
-        let got2 = lexical_identifiers_in_expressions("{ f({ a: 'literal' }) }");
-        assert!(got2.contains(&"f".to_string()), "{got2:?}");
-        assert!(got2.contains(&"a".to_string()), "{got2:?}");
-        assert!(
-            !got2.contains(&"literal".to_string()),
-            "string skipped: {got2:?}"
-        );
-    }
-
-    #[test]
-    fn lexical_identifiers_in_expressions_no_panic_on_unterminated_brace() {
-        // Defensive: an unterminated `{` ending mid-multibyte must not panic.
-        let _ = lexical_identifiers_in_expressions("{ x = \u{1F600}");
-        let _ = lexical_identifiers_in_expressions("{\u{30A2}");
     }
 
     #[test]
